Done lesson 12 home work. Firebase. MVP model
