+++ app/src/main/java/com/example/notes/presenter/OnDialogListener.java
@@ -1,4 +1,4 @@
-package com.example.notes;
+package com.example.notes.presenter;
 
 import android.widget.TextView;
