+++ app/src/main/java/com/example/notes/view/DialogHeaderDrawerFragment.java
@@ -1,4 +1,4 @@
-package com.example.notes;
+package com.example.notes.view;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -7,6 +7,9 @@ import android.view.ViewGroup;
 
 import androidx.annotation.Nullable;
 import androidx.fragment.app.DialogFragment;
+
+import com.example.notes.R;
+import com.example.notes.presenter.OnDialogListener;
 
 public class DialogHeaderDrawerFragment extends DialogFragment {
     private OnDialogListener dialogListener;
