+++ app/src/main/res/layout-land/fragment_note.xml
@@ -8,7 +8,7 @@
     android:backgroundTint="@color/white"
     android:clickable="true"
     android:orientation="vertical"
-    tools:context=".NoteFragment">
+    tools:context=".view.NoteFragment">
 
     <ScrollView
         android:id="@+id/scroll_list_names"
