+++ app/src/main/res/layout/fragment_date_picker.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    tools:context=".DatePickerFragment">
+    tools:context=".view.DatePickerFragment">
 
     <DatePicker
         android:id="@+id/datePicker"
