+++ app/src/main/res/layout/fragment_list_v2.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ListFragmentV2">
+    tools:context=".presenter.adapter.ListFragmentV2">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_view_lines"
