+++ app/src/main/res/layout/fragment_parent_date_time.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    tools:context=".ParentDateTimeFragment">
+    tools:context=".presenter.ParentDateTimeFragment">
 
     <com.google.android.material.tabs.TabLayout
         android:id="@+id/tabLayout"
