+++ app/src/main/res/layout/fragment_time_picker.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    tools:context=".TimePickerFragment">
+    tools:context=".view.TimePickerFragment">
 
     <TimePicker
         android:id="@+id/timePicker"
